Add Routes getStoriesByTags() and getStoriesByTagsAndTitle()

+++ EStory_Server/src/service/TagService.kt
@@ -13,8 +13,7 @@ import java.util.*
 
 class TagService {
 
-
-    private val tags = listOf(
+    private val tags = mutableListOf(
         "adventure",
         "Tragedy",
         "Comedy",
@@ -33,8 +32,10 @@ class TagService {
     private fun insertTag(name: String): Tag = transaction { TagEntity.new { this.name = name }.toDTO() }
 
     fun insertAllTags() {
-        transaction { TagsTable.deleteAll() }
-        tags.forEach { insertTag(it) }
+        if ( getAllTags().size != tags.size) {
+            transaction { TagsTable.deleteAll() }
+            tags.forEach { insertTag(it) }
+        }
     }
 
     fun mapStoryToTag(storyId: String, tagName: String): TaggedStory = transaction {
@@ -49,6 +50,24 @@ class TagService {
     fun findStoriesByTag(tagName: String): List<TaggedStory> = getAllTaggedStories().filter { it.tag.name == tagName }
     fun getAllTagsToStory(storyId: String): List<Tag> =
         getAllTaggedStories().filter { it.story.uuid == storyId }.map { it.tag }
+
+    fun getStoriesByTags(listOfTags: List<String>): List<Story> {
+
+        val temporarilyList = mutableListOf<TaggedStory>()
+
+        val allTaggedStories = getAllTaggedStories()
+        listOfTags.forEach { tagName ->
+            val filteredList =
+                allTaggedStories.filter { it.tag.name == tagName } // the list can have duplicated elements
+            temporarilyList.addAll(filteredList)
+        }
+
+        return temporarilyList.distinctBy { it.story.uuid }.map { it.story }
+    }
+
+
+    fun getStoriesByTagsAndTitle(listOfTags: List<String>, title: String): List<Story> =
+        getStoriesByTags(listOfTags).filter { it.storyTitle == title }
 
 
 }
